requirements.txt
flask
python-dotenv
langchain
langchain-groq
langchain-pinecone
streamlit
pinecone-client
PyPDF2
PyMuPDF
pillow
requests